requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.5
colorama==0.4.6
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-bidi==0.6.7
python-dotenv==1.2.2
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
